+++ CHANGELOG.md
@@ -27,6 +27,7 @@
     * Fixed `Inflector` methods when translation is missing `GRAV.INFLECTOR_*` translations
     * Fixed exception when changing parent of new page [grav-plugin-admin#2018](https://github.com/getgrav/grav-plugin-admin/issues/2018)
     * Fixed ordering issue with moving pages [grav-plugin-admin#2015](https://github.com/getgrav/grav-plugin-admin/issues/2015)
+    * Fixed Flex Pages cache not invalidating if saving an old `Page` object [#3152](https://github.com/getgrav/grav/issues/3152)
 
 # v1.7.0-rc.20
 ## 12/15/2020